+++ backend/api/src/handlers/auth/update-client-in-registry.ts
