+++ backend/api/src/handlers/step-functions/do-update-client.ts
