+++ backend/api/src/onboarding-table-item.ts
@@ -1,5 +1,5 @@
 export interface OnboardingTableItem {
-    pk: string,
-    sk: string,
+    pk: string;
+    sk: string;
     [propName: string]: any;
 }
